+++ src/handle/common.js
@@ -1,6 +1,5 @@
 import { getError } from '../error/main.js'
 import { exitProcess } from '../exit.js'
-import { getLevel } from '../level.js'
 import { isLimited } from '../limit.js'
 import { isRepeated } from '../repeat.js'
 
@@ -33,20 +32,8 @@ export const handleEvent = async function ({
     return
   }
 
-  await logEvent({ opts, name, event })
-
-  await exitProcess({ name, opts })
-}
-
-const logEvent = async function ({ opts, name, event }) {
-  const { error, mainValue } = getError({ name, event })
-
-  const level = getLevel({ opts, name, error })
-
-  if (level === 'silent') {
-    return
-  }
-
+  const error = getError({ name, event })
   // See `exit.js` on why we need to `await`
-  await opts.log(error, level, mainValue)
+  await opts.log(error, name)
+  await exitProcess({ name, opts })
 }