+++ src/limit.js
@@ -6,8 +6,7 @@ import { emitWarning } from 'process'
 //  - the `repeated` logic should prevent reaching the threshold
 //  - `previousEvents` might otherwise take too much memory and/or create a
 //    memory leak.
-//  - it prevents infinite recursions if `opts.log|level()` triggers itself an
-//    event.
+//  - it prevents infinite recursions if `opts.log()` triggers itself an event.
 //    The `repeated` logic should prevent it most of the times, but it can still
 //    happen when `[next]Value` is not an `Error` instance and contain dynamic
 //    content.