+++ src/repeat.js
@@ -5,8 +5,8 @@ import { inspect } from 'util'
 //  - it prevents creating too much CPU load or too many microtasks
 //  - it prevents creating too many logs, which can be expensive if logs are
 //    hosted remotely
-//  - it prevents infinite recursions if `opts.log|level()` triggers itself an
-//    event (while still reporting that event once)
+//  - it prevents infinite recursions if `opts.log()` triggers itself an event
+//    (while still reporting that event once)
 export const isRepeated = function ({ event, previousEvents }) {
   const fingerprint = getFingerprint({ event })