+++ test/helpers/level.js
@@ -1,1 +1,0 @@
-export const LEVELS = ['debug', 'info', 'warn', 'error']